Move the constants file out to the top-level
This should allow consumers to do `import { START_DATE, END_DATE } from 'react-dates/constants';` which is ideal
