+++ constants.js
@@ -1,0 +1,10 @@
+module.exports = {
+  DISPLAY_FORMAT: 'L',
+  ISO_FORMAT: 'YYYY-MM-DD',
+
+  START_DATE: 'startDate',
+  END_DATE: 'endDate',
+
+  HORIZONTAL_ORIENTATION: 'horizontal',
+  VERTICAL_ORIENTATION: 'vertical'
+};
