--- src/constants.js
@@ -1,8 +1,0 @@
-export const DISPLAY_FORMAT = 'L';
-export const ISO_FORMAT = 'YYYY-MM-DD';
-
-export const START_DATE = 'startDate';
-export const END_DATE = 'endDate';
-
-export const HORIZONTAL_ORIENTATION = 'horizontal';
-export const VERTICAL_ORIENTATION = 'vertical';
