+++ src/shapes/FocusedInputShape.js
@@ -3,6 +3,6 @@ import { PropTypes } from 'react';
 import {
   START_DATE,
   END_DATE,
-} from '../constants';
+} from '../../constants';
 
 export default PropTypes.oneOf([START_DATE, END_DATE]);
