+++ src/shapes/OrientationShape.js
@@ -3,6 +3,6 @@ import { PropTypes } from 'react';
 import {
   HORIZONTAL_ORIENTATION,
   VERTICAL_ORIENTATION,
-} from '../constants';
+} from '../../constants';
 
 export default PropTypes.oneOf([HORIZONTAL_ORIENTATION, VERTICAL_ORIENTATION]);
