+++ src/utils/toLocalizedDateString.js
@@ -2,7 +2,7 @@ import moment from 'moment';
 
 import toMomentObject from './toMomentObject';
 
-import { DISPLAY_FORMAT } from '../constants';
+import { DISPLAY_FORMAT } from '../../constants';
 
 export default function toLocalizedDateString(date, currentFormat) {
   const dateObj = moment.isMoment(date) ? date : toMomentObject(date, currentFormat);
