+++ src/utils/toMomentObject.js
@@ -1,6 +1,6 @@
 import moment from 'moment';
 
-import { DISPLAY_FORMAT, ISO_FORMAT } from '../constants';
+import { DISPLAY_FORMAT, ISO_FORMAT } from '../../constants';
 
 export default function toMomentObject(dateString, customFormat) {
   if (customFormat) {
